Simplified pagination and unified HttpService
